refactor: rename package ddo_auth to ddorang
