+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'ddo-auth'
+rootProject.name = 'ddorang'
