--- src/main/java/com/example/ddo_auth/DdoAuthApplication.java
@@ -1,13 +1,0 @@
-package com.example.ddo_auth;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DdoAuthApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DdoAuthApplication.class, args);
-	}
-
-}
