--- src/main/java/com/example/ddo_auth/auth/dto/SignupRequest.java
@@ -1,21 +1,0 @@
-package com.example.ddo_auth.auth.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SignupRequest {
-
-    @NotBlank(message = "이메일은 필수입니다.")
-    @Email(message = "이메일 형식이 올바르지 않습니다.")
-    private String email;
-
-    @NotBlank(message = "비밀번호는 필수입니다.")
-    private String password;
-
-    @NotBlank(message = "이름은 필수입니다.")
-    private String name;
-}
