--- src/main/java/com/example/ddo_auth/auth/repository/RefreshTokenRepository.java
@@ -1,13 +1,0 @@
-package com.example.ddo_auth.auth.repository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository {
-    void save(String email, String token);
-    Optional<String> findByEmail(String email);
-    void deleteByEmail(String email);
-    void     saveMapping(String token, String email);   // RT → email
-    void     deleteByToken(String token);
-    Optional<String> findEmailByToken(String token);
-}
-
