--- src/main/java/com/example/ddo_auth/common/ApiPaths.java
@@ -1,20 +1,0 @@
-package com.example.ddo_auth.common;
-
-/* 경로 상수화 */
-public interface ApiPaths {
-    /* 인증 관련
-    /* 공통 루트 */
-    String ROOT = "/api";
-
-    /* Auth(로컬 이메일/비밀번호) */
-    String AUTH = ROOT + "/auth";
-
-    /* OAuth(Google) */
-    String OAUTH = ROOT + "/oauth2";
-
-    /* Refresh Token*/
-    String TOKEN_REFRESH = AUTH + "/token/refresh";
-
-    /* Logout */
-    String TOKEN_LOGOUT = AUTH + "/token/logout";
-}
