+++ src/main/java/com/example/ddorang/DdoAuthApplication.java
@@ -1,0 +1,13 @@
+package com.example.ddorang;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DdoAuthApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DdoAuthApplication.class, args);
+	}
+
+}
