--- src/test/java/com/example/ddo_auth/DdoAuthApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.ddo_auth;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DdoAuthApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
