+++ src/test/java/com/example/ddorang/DdoAuthApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.ddorang;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DdoAuthApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
